[#1067] Fixed the realtime logging for a model training and packaging
We had two problems with realtime log streaming:
* MultiReader doesn't work as expected. So we decided to rewrite it on
sequence reading.
* In the client side we must not read response body in case of
streaming.

Moreover, cleanup some unused functions.
