+++ legion/operator/pkg/webserver/routes/v1/training/model_training_validation.go
@@ -64,8 +64,8 @@ var (
 		},
 	}
 	expectedConnectionDataTypes = map[v1alpha1.ConnectionType]bool{
-		connection.GcsType: true,
-		connection.S3Type:  true,
+		connection.GcsType:       true,
+		connection.S3Type:        true,
 		connection.AzureBlobType: true,
 	}
 )
